seperate racingcargame class
